[WIP] Introduce migration compatibility per database adapters
`ActiveRecord::Migration::Compatibility` handles database migration differences
per Active Record versions. It also manages database adapter differences
by adding `if adapter_name` statements, which is a little bit hard to maintain
and I wanted these if statements disappeared
by separating migration compatibility per database adapters as Active Record connection adapter does.

It will also allow 3rd party database adapters to implement
their own migration differences per Active Record version.

This commit message has [WIP] tag since there are still some TODO things:

* Remove `adapter_name != "Mysql2")` condition
  from `ActiveRecord::Migration::Compatibility::V5_0#create_table`
* Suppress `-- find_migration_compatibility(1.0)` message
from `test_unknown_migration_version_should_raise_an_argument_error`
